Refactor to clean-up code
