+++ src/api.js
@@ -1,7 +1,6 @@
 const fetchData = (zipCode) => {
   return (fetch(`https://api.openbrewerydb.org/breweries?by_postal=${zipCode}`)
   .then(response => {
-    console.log('response', response)
     if(response.ok) {
       return response.json()
     } else {
