+++ src/components/App/App.css
@@ -1,0 +1,28 @@
+
+body {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  width: 100vw;
+  background: url('../../assets/beer.png') no-repeat center center fixed; 
+  -webkit-background-size: cover;
+  -moz-background-size: cover;
+  -o-background-size: cover;
+  background-size: cover;
+}
+
+.App {
+  font-family: Verdana, Tahoma, sans-serif;
+  color: black;
+  font-style: 40px;
+  padding-bottom: 20px;
+  /* height: 100vh; */
+  text-align: center;
+}
+
+div {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  border: black solid 2px;
+}
