+++ src/components/Errors/error.css
@@ -2,7 +2,7 @@ div {
   height: 60%
 }
 
-.no-data {
+.error {
   /* width: fit-content; */
   height: 60%;
   background: url('../../assets/thirsty.gif') no-repeat center center; 
@@ -16,8 +16,3 @@ div {
   color: whitesmoke;
   font-family: none;
 }
-
-/* section {
-  display: flex;
-  flex-direction: column;
-} */
